Fixed email multi sender bug
The `sendEmail` function was `pass`ing on the first successful sent email and not sending to other recepients.

Also enhanced the email by adding a host name tag at the bottom so that we can track which process is sending emails easier.

Also used the full `email_copy_html` template (which for some reason was previously unused) and added some more logging.