Fixes batching logic
There was a flaw in the batching logic that prevented sensors from activating after a pause timeout. The big fix here is that this notification system no longer depends on the date times in the sensor names at all; when the recording happened doesn't matter, only whether we have sent a notification for the recording.

There was a large refactor involved in this change to more explicitly model the state of the sensors.

The unit tests were improved and more of the code is now covered by tests.

Closes #4